+++ tartufo/types.py
@@ -81,6 +81,7 @@ class GlobalOptions:
       values, while a value of 100 will detect only wholly random values.
     :param color: Enable or disable terminal color. If not provided (default),
       enabled if output is a terminal (TTY).
+    :param no_local_config: Disable local tartufo config from being used.
     """
 
     __slots__ = (
@@ -105,6 +106,7 @@ class GlobalOptions:
         "output_format",
         "entropy_sensitivity",
         "color",
+        "no_local_config",
     )
     rule_patterns: Tuple[Dict[str, str], ...]
     default_regexes: bool
@@ -127,6 +129,7 @@ class GlobalOptions:
     output_format: Optional[OutputFormat]
     entropy_sensitivity: int
     color: Optional[bool]
+    no_local_config: bool
 
 
 @dataclass